Adding PWA constraints to MICP controller

+++ mie2c/e2c.py
@@ -372,7 +372,7 @@ def compute_loss(x_dec, x_next_dec, x_next_pred_dec,
     KLD = distributions.kl_divergence(Qz,prior) + distributions.kl_divergence(Qz_next,prior)
 
     # ELBO
-    bound_loss = x_reconst_loss.add(x_next_reconst_loss).double().add(KLD)
-    trans_loss = distributions.kl_divergence(Qz_next_pred, Qz_next) # .add(x_next_pre_reconst_loss)
+    bound_loss = x_reconst_loss.add(x_next_reconst_loss).double().add(KLD).float()
+    trans_loss = distributions.kl_divergence(Qz_next_pred, Qz_next).float() # .add(x_next_pre_reconst_loss)
     
     return bound_loss.mean()/2, trans_loss.mean()
